fix: author id INT vs LONG

+++ src/services/author-catalog/src/models/author.rs
@@ -3,5 +3,5 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, sqlx::FromRow, Clone, Serialize, Deserialize)]
 pub struct Author {
     pub name: String,
-    pub author_id: i32,
+    pub author_id: i64,
 }
